Move logging & add logging cvar
