+++ soh/soh/resource/importer/scenecommand/EndMarkerFactory.h
@@ -13,6 +13,4 @@ class EndMarkerFactoryXML : public SceneCommandFactoryXMLV0 {
     std::shared_ptr<LUS::IResource>
     ReadResource(std::shared_ptr<LUS::ResourceInitData> initData, tinyxml2::XMLElement* reader) override;
 };
-
-void LogEndMarkerAsXML(std::shared_ptr<LUS::IResource> resource);
 } // namespace SOH
