+++ soh/soh/resource/importer/scenecommand/SetCollisionHeaderFactory.cpp
@@ -1,5 +1,6 @@
 #include "soh/resource/importer/scenecommand/SetCollisionHeaderFactory.h"
 #include "soh/resource/type/scenecommand/SetCollisionHeader.h"
+#include "soh/resource/logging/SceneCommandLoggers.h"
 #include "libultraship/libultraship.h"
 #include "spdlog/spdlog.h"
 
@@ -13,7 +14,9 @@ std::shared_ptr<LUS::IResource> SetCollisionHeaderFactory::ReadResource(std::sha
     setCollisionHeader->fileName = reader->ReadString();
     setCollisionHeader->collisionHeader = std::static_pointer_cast<CollisionHeader>(LUS::Context::GetInstance()->GetResourceManager()->LoadResourceProcess(setCollisionHeader->fileName.c_str()));
 
-    //LogSetCollisionHeaderAsXML(setCollisionHeader);
+    if (CVarGetInteger("gDebugResourceLogging", 0)) {
+        LogSetCollisionHeaderAsXML(setCollisionHeader);
+    }
 
     return setCollisionHeader;
 }
@@ -28,20 +31,5 @@ std::shared_ptr<LUS::IResource> SetCollisionHeaderFactoryXML::ReadResource(std::
     setCollisionHeader->collisionHeader = std::static_pointer_cast<CollisionHeader>(LUS::Context::GetInstance()->GetResourceManager()->LoadResourceProcess(setCollisionHeader->fileName.c_str()));
 
     return setCollisionHeader;
-}
-
-void LogSetCollisionHeaderAsXML(std::shared_ptr<LUS::IResource> resource) {
-    std::shared_ptr<SetCollisionHeader> setCollisionHeader = std::static_pointer_cast<SetCollisionHeader>(resource);
-
-    tinyxml2::XMLDocument doc;
-    tinyxml2::XMLElement* root = doc.NewElement("SetCollisionHeader");
-    doc.InsertFirstChild(root);
-
-    root->SetAttribute("FileName", setCollisionHeader->fileName.c_str());
-
-    tinyxml2::XMLPrinter printer;
-    doc.Accept(&printer);
-
-    SPDLOG_INFO("{}: {}", resource->GetInitData()->Path, printer.CStr());
 }
 } // namespace SOH
