+++ soh/soh/resource/importer/scenecommand/SetCsCameraFactory.cpp
@@ -1,5 +1,6 @@
 #include "soh/resource/importer/scenecommand/SetCsCameraFactory.h"
 #include "soh/resource/type/scenecommand/SetCsCamera.h"
+#include "soh/resource/logging/SceneCommandLoggers.h"
 #include "spdlog/spdlog.h"
 
 namespace SOH {
@@ -14,7 +15,9 @@ SetCsCameraFactory::ReadResource(std::shared_ptr<LUS::ResourceInitData> initData
 
     // OTRTODO: FINISH!
 
-    //LogCsCameraAsXML(setCsCamera);
+    if (CVarGetInteger("gDebugResourceLogging", 0)) {
+        LogCsCameraAsXML(setCsCamera);
+    }
 
     return setCsCamera;
 }
@@ -26,18 +29,5 @@ std::shared_ptr<LUS::IResource> SetCsCameraFactoryXML::ReadResource(std::shared_
     setCsCamera->cmdId = SceneCommandID::SetCsCamera;
 
     return setCsCamera;
-}
-
-void LogCsCameraAsXML(std::shared_ptr<LUS::IResource> resource) {
-    std::shared_ptr<SetCsCamera> setCsCamera = std::static_pointer_cast<SetCsCamera>(resource);
-
-    tinyxml2::XMLDocument doc;
-    tinyxml2::XMLElement* root = doc.NewElement("SetCsCamera");
-    doc.InsertFirstChild(root);
-
-    tinyxml2::XMLPrinter printer;
-    doc.Accept(&printer);
-
-    SPDLOG_INFO("{}: {}", resource->GetInitData()->Path, printer.CStr());
 }
 } // namespace SOH
