+++ soh/soh/resource/importer/scenecommand/SetCsCameraFactory.h
@@ -14,6 +14,4 @@ class SetCsCameraFactoryXML : public SceneCommandFactoryXMLV0 {
     std::shared_ptr<LUS::IResource>
     ReadResource(std::shared_ptr<LUS::ResourceInitData> initData, tinyxml2::XMLElement* reader) override;
 };
-
-void LogCsCameraAsXML(std::shared_ptr<LUS::IResource> resource);
 } // namespace SOH
