+++ soh/soh/resource/importer/scenecommand/SetLightListFactory.cpp
@@ -1,5 +1,6 @@
 #include "soh/resource/importer/scenecommand/SetLightListFactory.h"
 #include "soh/resource/type/scenecommand/SetLightList.h"
+#include "soh/resource/logging/SceneCommandLoggers.h"
 #include "spdlog/spdlog.h"
 
 namespace SOH {
@@ -30,7 +31,9 @@ SetLightListFactory::ReadResource(std::shared_ptr<LUS::ResourceInitData> initDat
         setLightList->lightList.push_back(light);
     }
 
-    //LogLightListAsXML(setLightList);
+    if (CVarGetInteger("gDebugResourceLogging", 0)) {
+        LogLightListAsXML(setLightList);
+    }
 
     return setLightList;
 }
@@ -74,41 +77,5 @@ std::shared_ptr<LUS::IResource> SetLightListFactoryXML::ReadResource(std::shared
     setLightList->numLights = setLightList->lightList.size();
 
     return setLightList;
-}
-
-void LogLightListAsXML(std::shared_ptr<LUS::IResource> resource) {
-    std::shared_ptr<SetLightList> setLightList = std::static_pointer_cast<SetLightList>(resource);
-
-    tinyxml2::XMLDocument doc;
-    tinyxml2::XMLElement* root = doc.NewElement("SetLightList");
-    doc.InsertFirstChild(root);
-
-    for (size_t i = 0; i < setLightList->numLights; i++) {
-        tinyxml2::XMLElement* light = doc.NewElement("LightInfo");
-        light->SetAttribute("Type", setLightList->lightList[i].type);
-        if (false/*setLightList->lightList[i].type == LIGHT_DIRECTIONAL*/) {
-            light->SetAttribute("X", setLightList->lightList[i].params.dir.x);
-            light->SetAttribute("Y", setLightList->lightList[i].params.dir.y);
-            light->SetAttribute("Z", setLightList->lightList[i].params.dir.z);
-            light->SetAttribute("ColorR", setLightList->lightList[i].params.dir.color[0]);
-            light->SetAttribute("ColorG", setLightList->lightList[i].params.dir.color[1]);
-            light->SetAttribute("ColorB", setLightList->lightList[i].params.dir.color[2]);
-        } else {
-            light->SetAttribute("X", setLightList->lightList[i].params.point.x);
-            light->SetAttribute("Y", setLightList->lightList[i].params.point.y);
-            light->SetAttribute("Z", setLightList->lightList[i].params.point.z);
-            light->SetAttribute("ColorR", setLightList->lightList[i].params.point.color[0]);
-            light->SetAttribute("ColorG", setLightList->lightList[i].params.point.color[1]);
-            light->SetAttribute("ColorB", setLightList->lightList[i].params.point.color[2]);
-            light->SetAttribute("DrawGlow", setLightList->lightList[i].params.point.drawGlow);
-            light->SetAttribute("Radius", setLightList->lightList[i].params.point.radius);
-        }
-        root->InsertEndChild(light);
-    }
-
-    tinyxml2::XMLPrinter printer;
-    doc.Accept(&printer);
-
-    SPDLOG_INFO("{}: {}", resource->GetInitData()->Path, printer.CStr());
 }
 } // namespace SOH
