+++ soh/soh/resource/importer/scenecommand/SetLightingSettingsFactory.h
@@ -14,6 +14,4 @@ class SetLightingSettingsFactoryXML : public SceneCommandFactoryXMLV0 {
     std::shared_ptr<LUS::IResource>
     ReadResource(std::shared_ptr<LUS::ResourceInitData> initData, tinyxml2::XMLElement* reader) override;
 };
-
-void LogLightingSettingsAsXML(std::shared_ptr<LUS::IResource> resource);
 } // namespace SOH
