+++ soh/soh/resource/importer/scenecommand/SetMeshFactory.h
@@ -14,6 +14,4 @@ class SetMeshFactoryXML : public SceneCommandFactoryXMLV0 {
     std::shared_ptr<LUS::IResource>
     ReadResource(std::shared_ptr<LUS::ResourceInitData> initData, tinyxml2::XMLElement* reader) override;
 };
-
-void LogMeshAsXML(std::shared_ptr<LUS::IResource> resource);
 } // namespace SOH
