+++ soh/soh/resource/importer/scenecommand/SetRoomBehaviorFactory.cpp
@@ -1,5 +1,6 @@
 #include "soh/resource/importer/scenecommand/SetRoomBehaviorFactory.h"
 #include "soh/resource/type/scenecommand/SetRoomBehavior.h"
+#include "soh/resource/logging/SceneCommandLoggers.h"
 #include "spdlog/spdlog.h"
 
 namespace SOH {
@@ -12,7 +13,9 @@ SetRoomBehaviorFactory::ReadResource(std::shared_ptr<LUS::ResourceInitData> init
     setRoomBehavior->roomBehavior.gameplayFlags = reader->ReadInt8();
     setRoomBehavior->roomBehavior.gameplayFlags2 = reader->ReadInt32();
 
-    //LogRoomBehaviorAsXML(setRoomBehavior);
+    if (CVarGetInteger("gDebugResourceLogging", 0)) {
+        LogRoomBehaviorAsXML(setRoomBehavior);
+    }
 
     return setRoomBehavior;
 }
@@ -27,21 +30,5 @@ std::shared_ptr<LUS::IResource> SetRoomBehaviorFactoryXML::ReadResource(std::sha
     setRoomBehavior->roomBehavior.gameplayFlags2 = reader->IntAttribute("GameplayFlags2");
 
     return setRoomBehavior;
-}
-
-void LogRoomBehaviorAsXML(std::shared_ptr<LUS::IResource> resource) {
-    std::shared_ptr<SetRoomBehavior> setRoomBehavior = std::static_pointer_cast<SetRoomBehavior>(resource);
-
-    tinyxml2::XMLDocument doc;
-    tinyxml2::XMLElement* root = doc.NewElement("SetRoomBehavior");
-    doc.InsertFirstChild(root);
-
-    root->SetAttribute("GameplayFlags1", setRoomBehavior->roomBehavior.gameplayFlags);
-    root->SetAttribute("GameplayFlags2", setRoomBehavior->roomBehavior.gameplayFlags2);
-
-    tinyxml2::XMLPrinter printer;
-    doc.Accept(&printer);
-
-    SPDLOG_INFO("{}: {}", resource->GetInitData()->Path, printer.CStr());
 }
 } // namespace SOH
