+++ soh/soh/resource/importer/scenecommand/SetSkyboxModifierFactory.cpp
@@ -1,5 +1,6 @@
 #include "soh/resource/importer/scenecommand/SetSkyboxModifierFactory.h"
 #include "soh/resource/type/scenecommand/SetSkyboxModifier.h"
+#include "soh/resource/logging/SceneCommandLoggers.h"
 #include "spdlog/spdlog.h"
 
 namespace SOH {
@@ -12,7 +13,9 @@ std::shared_ptr<LUS::IResource> SetSkyboxModifierFactory::ReadResource(std::shar
     setSkyboxModifier->modifier.skyboxDisabled = reader->ReadInt8();
     setSkyboxModifier->modifier.sunMoonDisabled = reader->ReadInt8();
 
-    //LogSkyboxModifierAsXML(setSkyboxModifier);
+    if (CVarGetInteger("gDebugResourceLogging", 0)) {
+        LogSkyboxModifierAsXML(setSkyboxModifier);
+    }
 
     return setSkyboxModifier;
 }
@@ -27,21 +30,5 @@ std::shared_ptr<LUS::IResource> SetSkyboxModifierFactoryXML::ReadResource(std::s
     setSkyboxModifier->modifier.sunMoonDisabled = reader->IntAttribute("SunMoonDisabled");
 
     return setSkyboxModifier;
-}
-
-void LogSkyboxModifierAsXML(std::shared_ptr<LUS::IResource> resource) {
-    std::shared_ptr<SetSkyboxModifier> setSkyboxModifier = std::static_pointer_cast<SetSkyboxModifier>(resource);
-
-    tinyxml2::XMLDocument doc;
-    tinyxml2::XMLElement* root = doc.NewElement("SetSkyboxModifier");
-    doc.InsertFirstChild(root);
-
-    root->SetAttribute("SkyboxDisabled", setSkyboxModifier->modifier.skyboxDisabled);
-    root->SetAttribute("SunMoonDisabled", setSkyboxModifier->modifier.sunMoonDisabled);
-
-    tinyxml2::XMLPrinter printer;
-    doc.Accept(&printer);
-
-    SPDLOG_INFO("{}: {}", resource->GetInitData()->Path, printer.CStr());
 }
 } // namespace SOH
