+++ soh/soh/resource/importer/scenecommand/SetSkyboxSettingsFactory.cpp
@@ -1,5 +1,6 @@
 #include "soh/resource/importer/scenecommand/SetSkyboxSettingsFactory.h"
 #include "soh/resource/type/scenecommand/SetSkyboxSettings.h"
+#include "soh/resource/logging/SceneCommandLoggers.h"
 #include "spdlog/spdlog.h"
 
 namespace SOH {
@@ -14,7 +15,9 @@ std::shared_ptr<LUS::IResource> SetSkyboxSettingsFactory::ReadResource(std::shar
     setSkyboxSettings->settings.weather = reader->ReadInt8();
     setSkyboxSettings->settings.indoors = reader->ReadInt8();
 
-    //LogSkyboxSettingsAsXML(setSkyboxSettings);
+    if (CVarGetInteger("gDebugResourceLogging", 0)) {
+        LogSkyboxSettingsAsXML(setSkyboxSettings);
+    }
 
     return setSkyboxSettings;
 }
@@ -31,23 +34,5 @@ std::shared_ptr<LUS::IResource> SetSkyboxSettingsFactoryXML::ReadResource(std::s
     setSkyboxSettings->settings.indoors = reader->IntAttribute("Indoors");
 
     return setSkyboxSettings;
-}
-
-void LogSkyboxSettingsAsXML(std::shared_ptr<LUS::IResource> resource) {
-    std::shared_ptr<SetSkyboxSettings> setSkyboxSettings = std::static_pointer_cast<SetSkyboxSettings>(resource);
-
-    tinyxml2::XMLDocument doc;
-    tinyxml2::XMLElement* root = doc.NewElement("SetSkyboxSettings");
-    doc.InsertFirstChild(root);
-
-    root->SetAttribute("Unknown", setSkyboxSettings->settings.unk);
-    root->SetAttribute("SkyboxId", setSkyboxSettings->settings.skyboxId);
-    root->SetAttribute("Weather", setSkyboxSettings->settings.weather);
-    root->SetAttribute("Indoors", setSkyboxSettings->settings.indoors);
-
-    tinyxml2::XMLPrinter printer;
-    doc.Accept(&printer);
-
-    SPDLOG_INFO("{}: {}", resource->GetInitData()->Path, printer.CStr());
 }
 } // namespace SOH
