+++ soh/soh/resource/importer/scenecommand/SetSkyboxSettingsFactory.h
@@ -14,6 +14,4 @@ class SetSkyboxSettingsFactoryXML : public SceneCommandFactoryXMLV0 {
     std::shared_ptr<LUS::IResource>
     ReadResource(std::shared_ptr<LUS::ResourceInitData> initData, tinyxml2::XMLElement* reader) override;
 };
-
-void LogSkyboxSettingsAsXML(std::shared_ptr<LUS::IResource> resource);
 } // namespace SOH
