+++ soh/soh/resource/importer/scenecommand/SetTransitionActorListFactory.cpp
@@ -1,5 +1,6 @@
 #include "soh/resource/importer/scenecommand/SetTransitionActorListFactory.h"
 #include "soh/resource/type/scenecommand/SetTransitionActorList.h"
+#include "soh/resource/logging/SceneCommandLoggers.h"
 #include "spdlog/spdlog.h"
 
 namespace SOH {
@@ -28,7 +29,9 @@ std::shared_ptr<LUS::IResource> SetTransitionActorListFactory::ReadResource(std:
         setTransitionActorList->transitionActorList.push_back(entry);
     }
 
-    //LogTransitionActorListAsXML(setTransitionActorList);
+    if (CVarGetInteger("gDebugResourceLogging", 0)) {
+        LogTransitionActorListAsXML(setTransitionActorList);
+    }
 
     return setTransitionActorList;
 }
@@ -64,33 +67,5 @@ std::shared_ptr<LUS::IResource> SetTransitionActorListFactoryXML::ReadResource(s
     setTransitionActorList->numTransitionActors = setTransitionActorList->transitionActorList.size();
 
     return setTransitionActorList;
-}
-
-void LogTransitionActorListAsXML(std::shared_ptr<LUS::IResource> resource) {
-    std::shared_ptr<SetTransitionActorList> setTransitionActorList = std::static_pointer_cast<SetTransitionActorList>(resource);
-
-    tinyxml2::XMLDocument doc;
-    tinyxml2::XMLElement* root = doc.NewElement("SetTransitionActorList");
-    doc.InsertFirstChild(root);
-
-    for (size_t i = 0; i < setTransitionActorList->numTransitionActors; i++) {
-        tinyxml2::XMLElement* entry = doc.NewElement("TransitionActorEntry");
-        entry->SetAttribute("FrontSideRoom", setTransitionActorList->transitionActorList[i].sides[0].room);
-        entry->SetAttribute("FrontSideEffects", setTransitionActorList->transitionActorList[i].sides[0].effects);
-        entry->SetAttribute("BackSideRoom", setTransitionActorList->transitionActorList[i].sides[1].room);
-        entry->SetAttribute("BackSideEffects", setTransitionActorList->transitionActorList[i].sides[1].effects);
-        entry->SetAttribute("Id", setTransitionActorList->transitionActorList[i].id);
-        entry->SetAttribute("PosX", setTransitionActorList->transitionActorList[i].pos.x);
-        entry->SetAttribute("PosY", setTransitionActorList->transitionActorList[i].pos.y);
-        entry->SetAttribute("PosZ", setTransitionActorList->transitionActorList[i].pos.z);
-        entry->SetAttribute("RotY", setTransitionActorList->transitionActorList[i].rotY);
-        entry->SetAttribute("Params", setTransitionActorList->transitionActorList[i].params);
-        root->InsertEndChild(entry);
-    }
-
-    tinyxml2::XMLPrinter printer;
-    doc.Accept(&printer);
-
-    SPDLOG_INFO("{}: {}", resource->GetInitData()->Path, printer.CStr());
 }
 } // namespace SOH
