+++ soh/soh/resource/logging/SceneCommandLoggers.h
@@ -1,0 +1,31 @@
+#include "Resource.h"
+
+namespace SOH {
+
+void LogEndMarkerAsXML(std::shared_ptr<LUS::IResource> resource);
+void LogActorListAsXML(std::shared_ptr<LUS::IResource> resource);
+void LogAlternateHeadersAsXML(std::shared_ptr<LUS::IResource> resource);
+void LogCameraSettingsAsXML(std::shared_ptr<LUS::IResource> resource);
+void LogSetCollisionHeaderAsXML(std::shared_ptr<LUS::IResource> resource);
+void LogCsCameraAsXML(std::shared_ptr<LUS::IResource> resource);
+void LogCutscenesAsXML(std::shared_ptr<LUS::IResource> resource);
+void LogEchoSettingsAsXML(std::shared_ptr<LUS::IResource> resource);
+void LogEntranceListAsXML(std::shared_ptr<LUS::IResource> resource);
+void LogExitListAsXML(std::shared_ptr<LUS::IResource> resource);
+void LogLightingSettingsAsXML(std::shared_ptr<LUS::IResource> resource);
+void LogLightListAsXML(std::shared_ptr<LUS::IResource> resource);
+void LogMeshAsXML(std::shared_ptr<LUS::IResource> resource);
+void LogObjectListAsXML(std::shared_ptr<LUS::IResource> resource);
+void LogPathwaysAsXML(std::shared_ptr<LUS::IResource> resource);
+void LogRoomBehaviorAsXML(std::shared_ptr<LUS::IResource> resource);
+void LogRoomListAsXML(std::shared_ptr<LUS::IResource> resource);
+void LogSkyboxModifierAsXML(std::shared_ptr<LUS::IResource> resource);
+void LogSkyboxSettingsAsXML(std::shared_ptr<LUS::IResource> resource);
+void LogSoundSettingsAsXML(std::shared_ptr<LUS::IResource> resource);
+void LogSpecialObjectsAsXML(std::shared_ptr<LUS::IResource> resource);
+void LogStartPositionListAsXML(std::shared_ptr<LUS::IResource> resource);
+void LogTimeSettingsAsXML(std::shared_ptr<LUS::IResource> resource);
+void LogTransitionActorListAsXML(std::shared_ptr<LUS::IResource> resource);
+void LogWindSettingsAsXML(std::shared_ptr<LUS::IResource> resource);
+
+} //namespace SOH
